1. Bring back analytic version one; count all log by log level only 2. Creating analytic version tow; count all log by log level, application name, and node id 3. Update database migration 4. Update config file example

+++ models/timber_wolf.go
@@ -5,5 +5,5 @@ type TimberWolf struct {
 	ApplicationName string `json:"app_name"`
 	NodeId          string `json:"node_id"`
 	LogLevel        string `json:"log_level"`
-	Counter         int64  `json:"counter"`
+	Counter         int    `json:"counter"`
 }
